Api users by vote with pagination

See merge request Krat12/p_sto_7!45

+++ src/main/java/com/javamentor/qa/platform/models/dto/user/UserDto.java
@@ -12,6 +12,6 @@ public class UserDto {
     private String fullName;
     private String imageLink;
     private String city;
-    private int reputation;
+    private Long reputation;
 
 }
